requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
filelock==3.17.0
Flask==3.1.0
frozenlist==1.4.1
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langsmith==0.1.129
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==11.0.0
pipdeptree==2.25.0
playwright==1.49.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.5.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.1
tenacity==8.5.0
textblob==0.19.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.13.0
